refactor: move non-xml-specific fns to utils
